+++ src/license.ts
@@ -2,10 +2,26 @@ import { FileBase } from './file';
 import * as fs from 'fs-extra';
 import { Project } from './project';
 
+export interface LicenseOptions {
+  /**
+   * Copyright owner
+   *
+   * @default ""
+   */
+  readonly copyrightOwner?: string;
+
+  /**
+   * Period of license (e.g. "1998-2023")
+   *
+   * @default - current year (e.g. "2020")
+   */
+  readonly copyrightPeriod?: string;
+}
+
 export class License extends FileBase {
   private readonly text: string;
 
-  constructor(project: Project, spdx: string) {
+  constructor(project: Project, spdx: string, options: LicenseOptions) {
     super(project, 'LICENSE');
 
     const textFile = `${__dirname}/../license-text/${spdx}.txt`;
@@ -13,7 +29,12 @@ export class License extends FileBase {
       throw new Error(`unsupported license ${spdx}`);
     }
 
-    this.text = fs.readFileSync(textFile, 'utf-8');
+    const years = options.copyrightPeriod ?? new Date().getFullYear().toString();
+    const owner = options.copyrightOwner ?? '';
+
+    this.text = fs.readFileSync(textFile, 'utf-8')
+      .replace('[yyyy]', years)
+      .replace('[name of copyright owner]', owner)
   }
 
   protected get data(): string {