+++ src/node-project.ts
@@ -70,6 +70,13 @@ export interface CommonOptions {
   readonly releaseWorkflow?: boolean;
 
   /**
+   * Branches which trigger a release.
+   *
+   * @default [ "master" ]
+   */
+  readonly releaseBranches?: string[];
+
+  /**
    * Workflow steps to use in order to bootstrap this repo.
    *
    * @default - [ { run: `npx projen${PROJEN_VERSION}` }, { run: 'yarn install --frozen-lockfile' } ]
@@ -112,20 +119,99 @@ export interface CommonOptions {
    * @default - same as `minNodeVersion`
    */
   readonly workflowNodeVersion?: string;
+
+  /**
+   * The dist-tag to use when releasing to npm.
+   *
+   * @default "latest"
+   */
+  readonly npmDistTag?: string;
+
+  /**
+   * License copyright owner.
+   *
+   * @default - defaults to the value of authorName or "" if `authorName` is undefined.
+   */
+  readonly copyrightOwner?: string;
+
+  /**
+   * The copyright years to put in the LICENSE file.
+   * @default - current year
+   */
+  readonly copyrightPeriod?: string;
 }
 
 export interface NodeProjectOptions extends ProjectOptions, CommonOptions {
+  /**
+   * This is the name of your package. It gets used in URLs, as an argument on the command line,
+   * and as the directory name inside node_modules.
+   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-name
+   */
   readonly name: string;
+
+  /**
+   * The description is just a string that helps people understand the purpose of the package.
+   * It can be used when searching for packages in a package manager as well.
+   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-description
+   */
   readonly description?: string;
+
+  /**
+   * The repository is the location where the actual code for your package lives.
+   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-repository
+   */
   readonly repository?: string;
+
+  /**
+   * If the package.json for your package is not in the root directory (for example if it is part of a monorepo),
+   * you can specify the directory in which it lives.
+   */
   readonly repositoryDirectory?: string;
+
+  /**
+   * Author's name
+   */
   readonly authorName?: string;
+
+  /**
+   * Author's e-mail
+   */
   readonly authorEmail?: string;
-  readonly homepage?: string;
+
+  /**
+   * Author's URL / Website
+   */
   readonly authorUrl?: string;
+
+  /**
+   * Author's Organization
+   */
+  readonly authorOrganization?: boolean;
+
+  /**
+   * Package's Homepage / Website
+   */
+  readonly homepage?: string;
+
+  /**
+   * License's SPDX identifier.
+   * See https://github.com/eladb/projen/tree/master/license-text for a list of supported licenses.
+   */
   readonly license?: string;
+
+  /**
+   * Package's Stability
+   */
   readonly stability?: string;
+
+  /**
+   * Additional entries to .gitignore
+   */
   readonly gitignore?: string[];
+
+  /**
+   * Additional entries to .npmignore
+   */
   readonly npmignore?: string[];
 }
 
@@ -156,6 +242,8 @@ export class NodeProject extends Project {
   public readonly minNodeVersion?: string;
   public readonly maxNodeVersion?: string;
 
+  protected readonly npmDistTag: string;
+
   constructor(options: NodeProjectOptions) {
     super(options);
 
@@ -171,6 +259,8 @@ export class NodeProject extends Project {
     if (this.maxNodeVersion) {
       nodeVersion += ` <= ${this.maxNodeVersion}`;
     }
+
+    this.npmDistTag = options.npmDistTag ?? 'latest';
 
     this.manifest = {
       '//': GENERATION_DISCLAIMER,
@@ -188,6 +278,7 @@ export class NodeProject extends Project {
         name: options.authorName,
         email: options.authorEmail,
         url: options.authorUrl,
+        organization: options.authorOrganization ?? false,
       },
       'homepage': options.homepage,
       'devDependencies': this.devDependencies,
@@ -208,13 +299,29 @@ export class NodeProject extends Project {
     this.addBundledDependencies(...options.bundledDependencies ?? []);
 
     this.npmignore = new IgnoreFile(this, '.npmignore');
-
     this.addDefaultGitIgnore();
+
+    if (options.gitignore?.length) {
+      this.gitignore.comment('custom gitignore entries')
+      for (const i of options.gitignore) {
+        this.gitignore.exclude(i);
+      }
+    }
+
+    if (options.npmignore?.length) {
+      this.npmignore.comment('custom npmignore entries')
+      for (const i of options.npmignore) {
+        this.npmignore.exclude(i);
+      }
+    }
 
     // set license and produce license file
     const license = options.license ?? 'Apache-2.0';
     this.manifest.license = license;
-    new License(this, license);
+    new License(this, license, {
+      copyrightOwner: options.copyrightOwner ?? options.authorName,
+      copyrightPeriod: options.copyrightPeriod,
+    });
 
     this.addScripts({ projen: `node ${PROJEN_RC} && yarn install` });
     this.addScripts({ 'projen:upgrade': 'yarn upgrade projen && yarn projen' });
@@ -250,8 +357,9 @@ export class NodeProject extends Project {
     }
 
     if (options.releaseWorkflow ?? true) {
+      const releaseBranches = options.releaseBranches ?? [ 'master' ];
       this.releaseWorkflow = new NodeBuildWorkflow(this, 'Release', {
-        trigger: { push: { branches: [ 'master' ] } },
+        trigger: { push: { branches: releaseBranches } },
         uploadArtifact: true,
         nodeVersion: options.workflowNodeVersion ?? this.minNodeVersion,
         bootstrapSteps: options.workflowBootstrapSteps,
@@ -278,6 +386,7 @@ export class NodeProject extends Project {
                 run: 'npx -p jsii-release jsii-release-npm',
                 env: {
                   NPM_TOKEN: '${{ secrets.NPM_TOKEN }}',
+                  NPM_DIST_TAG: this.npmDistTag,
                 },
               },
             ],